+++ bingus/src/app/api/post_comment/route.ts
@@ -1,6 +1,6 @@
 // api/post_comment
 import { NextRequest } from "next/server";
-import pool from "../pool";
+import pool from "../../../lib/pool";
 
 export async function GET(request: NextRequest) {
     try {
