+++ bingus/src/app/api/posts/route.ts
@@ -1,6 +1,6 @@
 // api/posts
 import { NextRequest } from "next/server";
-import pool from "../pool";
+import pool from "../../../lib/pool";
 
 export async function GET(request: NextRequest) {
     try {
