+++ bingus/src/app/api/user_settings/route.ts
@@ -1,6 +1,6 @@
 // api/user_settings
 import { NextRequest } from "next/server";
-import pool from "../pool";
+import pool from "../../../lib/pool";
 
 export async function GET(request: NextRequest) {
     try {
