+++ bingus/src/lib/pool.ts
@@ -4,11 +4,11 @@ import pg from "pg";
 const { Pool } = pg;
 
 const pool = new Pool({
-    user: "postgres",
-    password: "Bingus_LLC",
-    host: "bingus-db-1.c9ayqsiuu3wz.us-east-1.rds.amazonaws.com",
+    user: process.env.DB_USER,
+    password: process.env.DB_PASSWORD,
+    host: process.env.DB_HOST,
     port: 5432,
-    database: "bingus",
+    database: process.env.DB_NAME,
     ssl: {
         rejectUnauthorized: false,
     },
